api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]>=2.0.0
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
groq
python-dotenv
